Переделал тикеты
